40k: write test specs for all 40k files
Update 40k test writer to find all 40k files, and write test specs for them, rather than listing each individually.

This brings up the case of .rosz files. In some cases, there are both .ros and .rosz versions of a file; prefer the .ros and delete the rosz, if so.