40k: fix #77 by handling Prayers
Fix #77 by handling profiles with typeName="Prayers".

Funny enough, there's a TODO to handle unknown profile types, which is a better way to address this bug.

Note that tests are not updated because unit abilities are not tracked in the tests, but this should be fixed when the aforementioned TODO is fixed.