40k: merge identical units on reoder
For 40, generate a hash for each unit so we can identify identical datasheets when they are reordered, and collapse them into one (with a count).

The hash could in the future be used to uniquely identify roster files, allowing us to store order in localStorage and then retrieve on roster load.