40k: Move to TSX
Use TSX template for 40k. This includes an implementation of JSX's createElement to generate HTML via the same document.createElement calls we previously used.

The HTML output is almost identical (save for an extra "width: 30%;" in the last cell of a few rows). This means we continue a few silly patterns:
 - <td> doesn't need scope attribute (it's only for <th>)
 - there are a couple cases of unnecessarily nested divs