40k: move rest of 10th rendering to tsx
Move rest of 40k 10th rendering to tsx, namely removing explicit calls to document.createElement().

Common files, like for rendering options, are untouched. Those functions require passing in parent component and then adding DOM to that, whereas with the tsx code we're writing, we want functions to return DOM nodes that are then added to parent components by the caller.

In future, we could update those common files, but it's lo pri and I'm keen to leave them alone